(feat: cleanup) cleaned up the implementation of rust halo2 ffi
