+++ mopro-ffi/src/app_config/ios.rs
@@ -1,4 +1,3 @@
-use std::fmt::format;
 use std::path::{Path, PathBuf};
 use std::process::Command;
 use std::{fs, io};
@@ -7,8 +6,6 @@ use super::{cleanup_tmp_local, install_arch, mktemp_local};
 
 // Load environment variables that are specified by by xcode
 pub fn build(library_name: &str, manifest_dir: &str, bindings_dir: &str) {
-    let cwd = std::env::current_dir().unwrap();
-
     let build_dir = format!("{}/build", manifest_dir);
     let build_dir_path = Path::new(&build_dir);
     let bindings_dir_path = Path::new(&bindings_dir);
