+++ mopro-ffi/src/app_config/utils.rs
@@ -1,0 +1,67 @@
+use std::path::{Path, PathBuf};
+use std::process::Command;
+use std::{fs, io};
+use uuid::Uuid;
+
+pub fn build_release() -> io::Result<()> {
+    // Set up the command to run `cargo build --release`
+    let output = Command::new("cargo")
+        .arg("build")
+        .arg("--release")
+        .spawn()
+        .expect("Failed to spawn cargo build")
+        .wait()
+        .expect("cargo build errored");
+
+    if !output.success() {
+        return Err(io::Error::new(
+            io::ErrorKind::Other,
+            "Failed to build the release library".to_string(),
+        ));
+    }
+
+    Ok(())
+}
+
+fn tmp_local(build_path: &Path) -> PathBuf {
+    let tmp_path = build_path.join("tmp");
+    if let Ok(metadata) = fs::metadata(&tmp_path) {
+        if !metadata.is_dir() {
+            panic!("non-directory tmp");
+        }
+    } else {
+        fs::create_dir_all(&tmp_path).expect("Failed to create local tmpdir");
+    }
+    tmp_path
+}
+
+pub fn mktemp_local(build_path: &Path) -> PathBuf {
+    let dir = tmp_local(build_path).join(&Uuid::new_v4().to_string());
+    fs::create_dir(&dir).expect("Failed to create tmpdir");
+    dir
+}
+
+pub fn cleanup_tmp_local(build_path: &Path) {
+    fs::remove_dir_all(tmp_local(build_path)).expect("Failed to remove tmpdir");
+}
+
+pub fn install_ndk() {
+    Command::new("cargo")
+        .arg("install")
+        .arg("cargo-ndk")
+        .spawn()
+        .expect("Failed to spawn cargo, is it installed?")
+        .wait()
+        .expect("Failed to install cargo-ndk");
+}
+
+pub fn install_arch(arch: String) {
+    Command::new("rustup")
+        .arg("target")
+        .arg("add")
+        .arg(arch.clone())
+        .spawn()
+        .expect("Failed to spawn rustup, is it installed?")
+        .wait()
+        .expect(format!("Failed to install target architecture {}", arch).as_str());
+}
